autenticação

+++ routes/web.php
@@ -1,6 +1,7 @@
 <?php
 
 use App\Http\Controllers\ReservaController;
+use App\Http\Controllers\UserController;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -14,22 +15,23 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-Route::delete('/reservas/{id}', [ReservaController::class, 'destroy'])->name('reservas.destroy');
-Route::get('/reservas/adicionar', [ReservaController::class, 'adicionar'])->name('reservas.adicionar');
-Route::post('/reservas', [ReservaController::class, 'store'])->name('reservas.store');
-Route::get('/reservas/{id}', [ReservaController::class, 'edit'])->name('reservas.ver');
-Route::put('/reservas/{id}', [ReservaController::class, 'update'])->name('reservas.update');
-Route::get('/reservas', [ReservaController::class, 'reservas'])->name('reservas');
-Route::post('/config', [ReservaController::class, 'config'])->name('config');
-Route::get('/config', [ReservaController::class, 'viewConfig'])->name('viewConfig');
-Route::get('/', [ReservaController::class, 'index'])->name('index');
+Route::delete('/reservas/{id}', [ReservaController::class, 'destroy'])->name('reservas.destroy')->middleware('admin');
+Route::get('/reservas/adicionar', [ReservaController::class, 'adicionar'])->name('reservas.adicionar')->middleware('admin');
+Route::post('/reservas', [ReservaController::class, 'store'])->name('reservas.store')->middleware('admin');
+Route::get('/reservas/{id}', [ReservaController::class, 'edit'])->name('reservas.ver')->middleware('admin');
+Route::put('/reservas/{id}', [ReservaController::class, 'update'])->name('reservas.update')->middleware('admin');
+Route::get('/reservas', [ReservaController::class, 'reservas'])->name('reservas')->middleware('admin');
+Route::post('/config', [ReservaController::class, 'config'])->name('config')->middleware('admin');
+Route::get('/config', [ReservaController::class, 'viewConfig'])->name('viewConfig')->middleware('admin');
+Route::get('/', [ReservaController::class, 'index'])->name('index')->middleware('admin');
+Route::get('/perfil', [UserController::class, 'perfil'])->name('perfil')->middleware('admin');
+Route::put('/perfil', [UserController::class, 'update'])->name('update.user')->middleware('admin');
+
+Route::get('/login', [UserController::class, 'login'])->name('login');
+Route::post('/auth', [UserController::class, 'auth'])->name('auth');
+Route::get('/logout', [UserController::class, 'logout'])->name('logout');
 
 
-Route::get('/perfil', function () {
-    return view('perfil');
-})->name('perfil');
 
 
-Route::get('/login', function () {
-    return view('login');
-})->name('login');
+
